refactor: reorganize TypeScript type definitions into dedicated modules
Move type-only definitions from composables into src/types/ for better
separation of concerns. API types go to types/api.ts, domain types to
types/domain.ts, and existing UI types to types/ui.ts. The barrel
index.ts re-exports everything. Composables retain re-exports for
backward compatibility.

Closes #33
