+++ src/composables/useApi.ts
@@ -1,7 +1,9 @@
 import type { Reactive, Ref } from 'vue'
-import type { Error } from '@/types'
+import type { ApiResponse, Error } from '@/types'
 import { reactive, ref } from 'vue'
 import { transformFeatures } from '@/utils/feature-transform'
+
+export type { Action, Actions, ActionType, ActionTypeOptions, ApiLink, ApiLinkGroups, ApiResponse, IFeature, Reason, ReasonGeom, ReasonTags } from '@/types'
 
 /**
  * Interface representing the reactive state and methods for handling API interactions.
@@ -40,97 +42,6 @@ interface ApiComposable {
    * Resets the error state
    */
   resetError: () => void
-}
-
-type ObjectType = 'node' | 'way' | 'relation'
-
-export type ActionType = 'accept' | 'reject'
-
-export type ActionTypeOptions = Record<string, string | string[] | object>
-
-export type Action = [string, ActionType | null, ActionTypeOptions | null]
-
-export type Actions = Record<string, Action[]>
-
-export interface ReasonGeom {
-  max_distance?: number
-  min_distance?: number
-  score: number
-  reason: string
-}
-
-export interface ReasonTags {
-  score: number
-  reason: string
-}
-
-export interface Reason {
-  geom: ReasonGeom
-  tags: ReasonTags
-  conflate: string
-}
-
-interface Changeset {
-  id: number
-  created_at: string
-  closed_at: string
-  open: boolean
-  user: string
-  uid: number
-  minlat: number
-  minlon: number
-  maxlat: number
-  maxlon: number
-  comments_count: number
-  changes_count: number
-  tags: Record<string, string>
-}
-
-export type ApiLinkGroups = Record<number, ApiLink[]>
-
-/**
- * Interface representing a link in the API metadata.
- * A link typically points to a feature.
- */
-export interface ApiLink {
-  action: ActionType
-  before?: number
-  after?: number
-  diff_attribs?: Actions
-  diff_tags?: Actions
-  conflation_reason: Reason
-}
-
-export interface IFeature extends GeoJSON.Feature {
-  id: number
-  properties: {
-    objtype: ObjectType
-    id: number
-    geom_distance: number | null
-    geom: boolean
-    deleted: boolean
-    links: number
-    members?: null
-    version: number
-    username: string
-    created: string
-    tags: Record<string, string>
-    is_before?: boolean
-    is_after?: boolean
-    is_new?: boolean
-  }
-}
-
-/**
- * Interface representing the API response.
- * Extends `GeoJSON.FeatureCollection` to represent geographic data and includes additional metadata.
- */
-export interface ApiResponse extends GeoJSON.FeatureCollection {
-  features: IFeature[]
-  metadata: {
-    links: ApiLinkGroups
-    changesets: Changeset[]
-  }
 }
 
 /**
