+++ src/types/index.ts
@@ -1,16 +1,3 @@
-export interface FormData {
-  dateStart: string
-  dateEnd?: string
-  bbox: string
-}
-
-export interface Preset extends FormData {
-  title: string
-}
-
-export type ErrorType = 'error' | 'warning' | 'success' | 'info'
-
-export interface Error {
-  message?: string
-  type: ErrorType
-}
+export type * from './api'
+export type * from './domain'
+export * from './ui'
